% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeR-package.R
\docType{package}
\name{AnalyzeR-package}
\alias{AnalyzeR}
\alias{AnalyzeR-package}
\title{AnalyzeR: Image Analyzsis Package}
\description{
A package for analyzing and quantifying staining in images. The functions provide a workflow for image analysis 'analyze_image()' processes a single image with a customizable threshold, 'group_loaded_images()' organizes images for batch processing, and 'batch_analyze_images()' analyzes all grouped images using the same threshold. The package calculates the percentage difference between stained and blank images, providing a quantitative measure of protein expression and summary statistics across image batches.
}
\author{
\strong{Maintainer}: Megan Hall \email{megan.hall@utexas.edu}

Authors:
\itemize{
  \item Kathryn Rhodes \email{kathryn.rhodes@utexas.edu}
  \item Patrick Mireles \email{pr_mireles@utexas.edu}
}

}
\keyword{internal}
