% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normAnalyzer.R
\name{anderson.test}
\alias{anderson.test}
\title{Anderson Darling test}
\usage{
anderson.test(x)
}
\arguments{
\item{x}{a numeric vector. Length must be greater than 7.
Missing values are allowed.}
}
\value{
A list with following elements:
\describe{
 \item{statistic}{the value of Anderson-Darling test statistic}
 \item{p.value}{p-value of the test}
 \item{method}{Test name}
 \item{data.name}{Vector name}
}
}
\description{
\code{anderson.test} performs Anderson-Darling test
}
\details{
Performs the Anderson-Darling test for  the composite hypothesis of normality,
see e.g. Thode (2002, Sec. 5.1.4).
}
\examples{
anderson.test(mtcars$mpg)

}
\seealso{
\code{\link{norm_test_fun}}
}
