% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{association}
\alias{association}
\title{Find association between variables}
\usage{
association(
  tb,
  categorical = NULL,
  method1 = c("auto", "pearson", "kendall", "spearman"),
  method3 = "auto",
  normality_test_method = c("shapiro", "ad"),
  use = "everything",
  ...
)
}
\arguments{
\item{tb}{tabular data}

\item{categorical}{a vector specifying the names of categorical (character,
factor) columns}

\item{method1}{method for association between continuous-continuous
variables. values can be \code{"auto", "pearson",  "kendall", "spearman"}.
See details for more information.}

\item{method3}{method for association between continuous-categorical
variables. Values can be \code{"auto", "parametric", "non-parametric"}.
See details of \code{\link{CQassociation}} for more information.
Parametric does t-test while non-parametric
does 'Mann-Whitney’ test.}

\item{normality_test_method}{method for normality test for a variable.
See details for more information. 'shapiro' or 'ad'}

\item{use}{an optional character string giving a method for computing
association in the presence of missing values. This must be (complete or an
abbreviation of) one of the strings "everything", "all.obs",
"complete.obs", "na.or.complete", or "pairwise.complete.obs". If use is
"everything", NAs will propagate conceptually, i.e., a resulting value will
be NA whenever one of its contributing observations is NA. If use is
"all.obs", then the presence of missing observations will produce an error.
If use is "complete.obs" then missing values are handled by casewise
deletion (and if there are no complete cases, that gives an error).
"na.or.complete" is the same unless there are no complete cases, that gives
NA}

\item{...}{other parameters passed to \code{cor}, \code{CCassociation},
\code{CQassociation} and \code{QQassociation}}
}
\value{
A list of three tables:
\describe{
 \item{continuous_corr}{correlation among all the continuous variables}
 \item{continuous_pvalue}{Table containing p-value for the correlation test}
 \item{categorical_cramers}{Cramer's V value among all the categorical variables}
 \item{categorical_pvalue}{Chi Sq test p-value}
 \item{continuous_categorical}{association value among continuous and categorical variables}
}
}
\description{
\code{association} finds association among all the variables in the data.
}
\details{
This function calculates association value in three categoris -
\itemize{
 \item between continuous variables (using \code{CCassociation} function)
 \item between categorical variables (using \code{QQassociation} function)
 \item between continuous and categorical variables (using \code{CQassociation}
  function)
}
For more details, look at the individual documentaion of
\code{\link{CCassociation}}, \code{\link{QQassociation}},
\code{\link{CQassociation}}
}
\examples{
tb <- mtcars
tb$cyl <- as.factor(tb$cyl)
tb$vs  <- as.factor(tb$vs)
association(tb, categorical = c("cyl", "vs"))
rm(tb)

}
\seealso{
\code{\link{CCassociation}} for Correlation between Continuous variables,
\code{\link{QQassociation}} for Association between Categorical variables,
\code{\link{CQassociation}} for Association between Continuous-Categorical variables
}
