% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCassociation.R
\name{CCassociation}
\alias{CCassociation}
\title{Association (Correlation) between Continous (numeric) Variables}
\usage{
CCassociation(numtb, use = "everything", norm_test_all = NULL)
}
\arguments{
\item{numtb}{a data frame with all the numerical columns. This should
have atleast two columns}

\item{use}{an optional character string giving a method for computing
association in the presence of missing values. This must be (complete or an
abbreviation of) one of the strings "everything", "all.obs",
"complete.obs", "na.or.complete", or "pairwise.complete.obs". If use is
"everything", NAs will propagate conceptually, i.e., a resulting value will
be NA whenever one of its contributing observations is NA. If use is
"all.obs", then the presence of missing observations will produce an error.
If use is "complete.obs" then missing values are handled by casewise
deletion (and if there are no complete cases, that gives an error).
"na.or.complete" is the same unless there are no complete cases, that gives
NA}

\item{norm_test_all}{a logical named vector with TRUE when the
column follows normality assumption, otherwise FALSE. This should have
TRUE or FALSE for all the columns present in \code{numtb}}
}
\value{
a table of correleations among the variables with number or
rows and column same as the number of columns in \code{numtb}.
}
\description{
\code{CCassociation} finds correlation between all the variables in data
with only numeric columns
}
\details{
This function calls \code{cor} function to calculate the correlation values.
The difference is that this doesn't take method as parameter, instead it
decides the methods iteself using normality tests. If the variables satisfy the
assumption of Pearson correlation, then pearson correlation is calculated.
Otherwise Spearman is calculated. To learn more, check the
\code{\link[stats]{cor}}
}
\examples{
CCassociation(mtcars)

# with norm_test_all
norm_test_all <- rep(TRUE, ncol(mtcars))
names(norm_test_all) <- colnames(mtcars)
CCassociation(mtcars, use = "complete.obs", norm_test_all = norm_test_all)
rm(norm_test_all)

}
\seealso{
\code{\link{association}} for association between any type of variables,
\code{\link{QQassociation}} for Association between Categorical variables,
\code{\link{CQassociation}} for Association between Continuous-Categorical variables
}
