% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_word_frequency.R
\name{check_word_frequency}
\alias{check_word_frequency}
\title{get count for a specific word of your word_amount_ngrams output}
\usage{
check_word_frequency(summary, word)
}
\arguments{
\item{summary}{A list given by word_amount_ngrams function}

\item{word}{the specific word that you want the frequency of}
}
\value{
A integer
}
\description{
get count for a specific word of your word_amount_ngrams output
}
\examples{
text <- "this day is a great day to talk about the next day"
summary <- word_amount_ngrams(text, 1)
check_word_frequency(summary, word = "day")
}
