% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_funcs.R
\name{Cx}
\alias{Cx}
\title{Plots for Continuous independent variables}
\usage{
Cx(dat, xname, binwidth = NULL, inc.density = T, ...)
}
\arguments{
\item{dat}{a data.frame with only one column}

\item{xname}{name of independent (x) variable in \code{dat}}

\item{binwidth}{for the histograms (extra parameters can be added like this)}

\item{inc.density}{Binary. True to include the density plot on histogram}

\item{...}{required}
}
\value{
a grob of plot
}
\description{
This function is used by [plottr()] when independent
variable is continuous. This function can be used as a template to
define a custom function.
}
