% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_funcs.R
\name{CxCy}
\alias{CxCy}
\title{Plots for Continuous independent and dependent variables}
\usage{
CxCy(dat, xname, yname, ...)
}
\arguments{
\item{dat}{a data.frame with two columns (including the dependent)}

\item{xname}{name of independent (x) variable in \code{dat}}

\item{yname}{name of dependent (y) variable in \code{dat}}

\item{...}{required}
}
\value{
a grob of plot
}
\description{
This function is used by [plottr()] when both the dependent and independent
variables are continuous. This function can be used as a template to
define a custom function
}
