% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_symbol.mesa.R
\name{examine_symbol.mesa}
\alias{examine_symbol.mesa}
\title{Gets the adjusted variable plot for mtDNA vs transcripts for people who remain after filtering}
\usage{
examine_symbol.mesa(lm_data, symbol, gene_key, omit.outlier = T,
  correct_for = "as.factor(lm_data$SEX) + as.numeric(lm_data$AGE) + as.numeric(lm_data$RACE) +\\n                  lm_data$PC1 + lm_data$PC2 + lm_data$PC3 + lm_data$PC4 + lm_data$PC5 + lm_data$PC6 + lm_data$PC7 + lm_data$PC8 + lm_data$PC9 + lm_data$PC10 + lm_data$Genotyping.PC1 + lm_data$Genotyping.PC2 + lm_data$Genotyping.PC3",
  col = NA)
}
\arguments{
\item{lm_data}{lm_full you want to examine the transcript in}

\item{symbol}{a string with the gene symbol you want to look at}

\item{gene_key}{a gene key with symbol -> transcript information (tpms_gene_key.rds)}

\item{omit.outlier}{T/F whether you want to omit outliers > 3SD away from the mean}

\item{correct_for}{a string that says what you need to correct transcripts for}

\item{col}{A string, what you want to color the adjusted variable plot with}
}
\value{
Transcripts for those individuals
}
\description{
Gets the adjusted variable plot for mtDNA vs transcripts for people who remain after filtering
}
\examples{
examine_transcript(blood_full, 'POLG', omit.outlier = T, correct_for = 'as.factor(lm_data$SEX)', col = 'esoph_day')
}
