% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Generic explain}
\usage{
explain(X, xname = NULL, ...)
}
\arguments{
\item{X}{a vector}

\item{xname}{name of the data to be printed. If missing then the
name of variable passed as \code{X} will be used}

\item{...}{other parameters required for other methods of \code{explain}
To see the parameters for numeric methods, use \code{?explain.numeric} and
similarly for other methods (\code{?explain.factor} etc.)}
}
\value{
Prints the information on the console. For print information for
the individual methods, see their documentation. Returns nothing.
}
\description{
Generic function for printing the details of a vector. Based on
the data type of vector, this calls the appropriate method.
}
\details{
Current methods for \code{explain} are for numeric, integer, character
and factor vectors. To get the list of all available methods type the
command \code{methods(explain)}.
}
\examples{
# for numeric
explain(mtcars$mpg) # same as explain.numeric(mtcars$mpg)
# for factor
explain(as.factor(mtcars$cyl)) # same as explain.factor(as.factor(mtcars$cyl))

}
