% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainer.R
\name{explainer.data.frame}
\alias{explainer.data.frame}
\title{Show details of the data frame}
\usage{
\method{explainer}{data.frame}(X, xname = NULL, ...)
}
\arguments{
\item{X}{A data.frame}

\item{xname}{variable name}

\item{...}{parameters for explainer for other classes}
}
\value{
Prints details of the dataset which includes: dataset name,
type, number of columns, rows and unique rows. Also prints output of
\code{explainer} for all the columns. Returns nothing.
}
\description{
\code{explainer} shows detail of all the columns of the data
}
\details{
This function uses \code{explainer} on each column.
}
\examples{
explainer(mtcars)

}
