% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explainer}
\alias{explainer}
\title{Show details of the data frame}
\usage{
explainer(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
Prints summary of the dataset
}
\description{
Show details of the data frame
}
\examples{
explainer(mtcars)
}
