% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_steps.R
\name{filtering_steps}
\alias{filtering_steps}
\title{Filters out people who had incomplete RNAseq downloads from GTEx to recount}
\usage{
filtering_steps(with_gtex)
}
\arguments{
\item{with_gtex}{the scaled rse gene counts (along with the GTEx ID info)}
}
\value{
The data frame, after removing people who had incomplete RNAseq downloads
}
\description{
Filters out people who had incomplete RNAseq downloads from GTEx to recount
}
\examples{

}
