% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{Finds PCA outliers - greater than 4 SD from the mean of the first 10 PCs.}
\usage{
find_outliers(pca)
}
\description{
Finds PCA outliers - greater than 4 SD from the mean of the first 10 PCs.
}
