% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.vars.R
\name{get.vars}
\alias{get.vars}
\title{Function that will get the variables you need to run all the lms in parallel}
\usage{
get.vars(lm_full, permutevar = "mtDNA_adjust_AGE",
  covnames = c(paste0("PC", 1:10), paste0("Genotyping.PC", 1:3), "sex",
  "COHORT"))
}
\arguments{
\item{lm_full}{A data frame with all variables.  Column = genes, covariates, etc. Rows = observations (in this case, individuals)}

\item{permutevar}{The main variable, as a string (for my case, mtDNA-CN)}

\item{covnames}{Names of the covariates to be included in the regression, a vector of strings.}
}
\value{
a list containing tx_expr, cov, gene.ids, and SCORE, all of which are to be fed into the function "run.all.lms"
}
\description{
Function that will get the variables you need to run all the lms in parallel
}
\examples{
my.list <- get.vars.runall.lms(lm_full, permutevar = 'mtDNA_adjust_AGE')
}
