% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transcripts.R
\name{get_transcripts}
\alias{get_transcripts}
\title{Gets the transcripts for people who remain after filtering}
\usage{
get_transcripts(counts1, with_gtex)
}
\arguments{
\item{counts1}{data frame containing the scaled counts from rse_gene (assay(rse_gene))}

\item{with_gtex}{data frame with GTEx IDs}
}
\value{
Transcripts for those individuals
}
\description{
Gets the transcripts for people who remain after filtering
}
