% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getid.R
\name{getid}
\alias{getid}
\title{Adds GTEx ID information to RNAseq data from recount}
\usage{
getid(scaled_rse)
}
\arguments{
\item{scaled_rse}{the scaled rse gene counts after it's been subsetted for tissue detail.}
}
\value{
Returns the scaled rse gene counts with GTEx ID information
}
\description{
Adds GTEx ID information to RNAseq data from recount
}
\examples{
pituitary <- subset(as.data.frame(colData(scaled)), smtsd== "Pituitary" & smafrze == 'USE ME')
pituitary <- getid(pituitary)
}
