% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmt.R
\name{getmt}
\alias{getmt}
\title{Load and merge mtDNA-CN info}
\usage{
getmt(transcripts_t, mt_frame = mtDNA_CN_with_pheno)
}
\arguments{
\item{transcripts_t}{scaled rse gene counts}

\item{mt_frame}{the data frame containing mtDNA-CN and read depth information}
}
\value{
Returns the data frame with mtDNA-CN info
}
\description{
Load and merge mtDNA-CN info
}
