% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_loaded_images.R
\name{group_loaded_images}
\alias{group_loaded_images}
\title{Group and Process Loaded Images}
\usage{
group_loaded_images(image_objects, processor_function, ...)
}
\arguments{
\item{image_objects}{A list of pre-loaded image objects to process}

\item{processor_function}{A function that will be applied to the image objects}

\item{...}{Additional arguments to pass to the processor function}
}
\value{
The result from the processor function
}
\description{
Takes a list of pre-loaded image objects and applies a specified processor function to them.
}
\examples{
\dontrun{
# Assuming img1, img2, and img3 are already loaded image objects
image_list <- list(img1, img2, img3)
result <- group_loaded_images(image_list, process_loaded_images, output_format = "png")
}
}
\keyword{images}
\keyword{processing}
