% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_pca_outliers.R
\name{identify_pca_outliers}
\alias{identify_pca_outliers}
\title{Filters transcripts based on expression levels.  Also filters out individuals who are PCA outliers}
\usage{
identify_pca_outliers(transcripts_t, name)
}
\description{
Filters transcripts based on expression levels.  Also filters out individuals who are PCA outliers
}
