% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv.norm.transform.R
\name{inv.norm.transform}
\alias{inv.norm.transform}
\title{Code for an inverse normal transformation, from Ryan}
\usage{
inv.norm.transform(mtDNA.resid.15nat)
}
\arguments{
\item{mtDNA.resid.15nat}{The variable that you'd like to inverse normal transform}
}
\value{
The same value, but after an inverse normal transformation
}
\description{
Code for an inverse normal transformation, from Ryan
}
\examples{
mtDNA.resid.15nat.inv <- inv.norm.transform(mtDNA.resid.15nat)
}
