% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{linedivider}
\alias{linedivider}
\title{Draws a horizontal line on console}
\usage{
linedivider(consolewidth)
}
\arguments{
\item{consolewidth}{a integer}
}
\value{
Prints a horizontal line of width 'consolewidth'
}
\description{
Draws a horizontal line on console
}
\examples{
linedivider(20)
}
