% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look_pca.R
\name{look_pca}
\alias{look_pca}
\title{Performs PCA, and tells you what percent of the variance each PC is.}
\usage{
look_pca(no.outlier)
}
\description{
Performs PCA, and tells you what percent of the variance each PC is.
}
