% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.forestplot.R
\name{make.forestplot}
\alias{make.forestplot}
\title{Makes forestplots for genes of interest across tissues}
\usage{
make.forestplot(gene, combined)
}
\arguments{
\item{gene}{gene you want to look at}

\item{combined}{dataframe with genes + effect sizes + standard errors for each tissue}
}
\value{
writes out a forestplot to a pdf
}
\description{
Makes forestplots for genes of interest across tissues
}
\examples{
make.forestplot('YBX1P10', combined)
}
