% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.subjids.R
\name{make.subjids}
\alias{make.subjids}
\title{Function that will return GTEx subject IDs when given sample IDs}
\usage{
make.subjids(sampids)
}
\arguments{
\item{x}{GTEx sample IDs}
}
\value{
GTEx subject-level IDs
}
\description{
Function that will return GTEx subject IDs when given sample IDs
}
\examples{
blood.runs$submitted_subject_id <- make.subjids(blood.runs$sampid)
}
