% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transcripts.R
\name{make_transcripts}
\alias{make_transcripts}
\title{Filter the transcripts to match the individuals that have been filtered out}
\usage{
make_transcripts(transcripts_t, with.mt.complete)
}
\description{
Filter the transcripts to match the individuals that have been filtered out
}
