% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.for.cutoff.R
\name{perm.for.cutoff}
\alias{perm.for.cutoff}
\title{Function that permutes the dataset and finds an appropriate cutoff value.}
\usage{
perm.for.cutoff(permutations = 100, lm_full)
}
\arguments{
\item{permutations}{The number of times you want to permute the data}

\item{lm_full}{The dataset you want to test}
}
\value{
A 95% cutoff value, also a data frame with all the p-values from each permutation (to make the overlaid p-val qqplot)
}
\description{
Function that permutes the dataset and finds an appropriate cutoff value.
}
\examples{
perm.results <- perm.for.cutoff(permutations = 5, subcut_full)
}
