% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.for.cutoff2.R
\name{perm.for.cutoff2}
\alias{perm.for.cutoff2}
\title{Function that permutes the dataset and finds an appropriate cutoff value.  (Compatible with Vamsee's run.all.lms)}
\usage{
perm.for.cutoff2(permutations = 100, tx_expr, cov, gene.ids, SCORE,
  num.cores = 10)
}
\arguments{
\item{tx_expr}{Expression matrix in form: [genes x samples]. Will be converted to a list(!) of gene-expr vectors (if not input as list)}

\item{cov}{Regression covariates [cov x samples]}

\item{gene.ids}{Character vector of gene IDs, corresponding to rows in `tx_expr` [genes x samples]}

\item{SCORE}{Main covariate to be permuted (not included in `cov`)}

\item{num.cores}{The number of cores you would like to use}
}
\value{
A 95% cutoff value, also a data frame with all the p-values from each permutation (to make the overlaid p-val qqplot)
}
\description{
Function that permutes the dataset and finds an appropriate cutoff value.  (Compatible with Vamsee's run.all.lms)
}
\examples{
perm.results <- perm.for.cutoff2(permutations = 100, my.list[[1]], my.list[[2]], my.list[[3]], my.list[[4]])
}
