% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottr.R
\name{plottr}
\alias{plottr}
\title{Exploratory plots of a variable}
\usage{
plottr(
  x,
  y = NULL,
  xclass = NULL,
  yclass = NULL,
  xname = "x",
  yname = "y",
  binwidth.x = NULL,
  binwidth.y = NULL,
  callasfactor = 1,
  inc.density = FALSE,
  ...
)
}
\arguments{
\item{x}{the vector of which visualization is needed}

\item{y}{the repsonse variable (if available, else NULL). Default is NULL}

\item{xclass}{class of x, if \code{xclass = NULL} then function
assigns the class \code{(factor/numeric)}}

\item{yclass}{class of x, if \code{y} is not NULL than this should
be set as either \code{factor} or \code{numeric}}

\item{xname, yname}{name of x and y. Can be the column name.}

\item{binwidth.x, binwidth.y}{binwidth for \code{x} and \code{y}
used for generating the histogram.
If there are \code{NULL}, then ggplot2 assigns the value
on its own. Look at \code{binwidth} of ggplot for more information}

\item{callasfactor}{minimum unique values needed for \code{x} to be
considered as numeric. See details for more information}

\item{inc.density}{(logical) Whether density plot is needed along
with histogram.}

\item{...}{Extra parameters to pass to ggplot}
}
\value{
This function returns a grob of class 'analyzerPlot' with
exploratory plots. Based on class of \code{x} and \code{y}, this
generates different plots including histogram, density, box,
crosstab heatmap etc. To view the plot, use \code{plot()}
}
\description{
\code{plottr} returns a grob with multiple plots explaining the variable
}
\details{
This is a function which helps in understanding the data through multiple
visualizations. This works either for a single \code{x} variable
or for a combination of predictor \code{x} and repsonse \code{y} variables.
Based on class of \code{x} and \code{y} different types of
plots are automatically generated.

Defining the class of \code{x} and \code{y}: If \code{y} is not NULL,
then \code{yclass} has to be passed (which can be 'factor' for
classification type problem, or 'numeric' for regression). \code{xclass}
can be set as NULL, in such case the function assigns a class
based on the contents. If \code{x} is factor/character type, then
\code{xclass = 'factor'}, else if \code{x} is numeric with
number of unique values less than \strong{callasfactor} parameter value,
then \code{xclass = 'factor'}, else \code{xclass = 'numeric'}

Types of plots: If the \code{y} is \code{NULL}, then histogram with
density (when \code{inc.density = TRUE}) is generated for numeric \code{x}.
Boxplot is also shown in the same histogram using color and vertical lines.
For factor \code{x}, a pie chart showing the distribution. If \code{y}
is not \code{NULL}, then additional plots are added:

\itemize{
 \item \strong{factor \code{x}, factor \code{y}}: Crosstab with heatmap
 \item \strong{factor \code{x}, numeric \code{y}}: histogram and
 boxplot of \code{y} for different values of \code{x}
 \item \strong{numeric \code{x}, factor \code{y}}: histogram and
 boxplot of \code{x} for different values of \code{y}
 \item \strong{numeric \code{x}, numeric \code{y}}: Scatter plot
 of \code{x} and \code{y} with rug plot included
}
}
\examples{
# simple plot with numerix x
p1 <- plottr(x = mtcars$mpg)
plot(p1)
# plot with numerix x and numeric y
p1 <- plottr(x = mtcars$mpg, y = mtcars$disp, xclass = "numeric",
             yclass = "numeric", xname = "mpg", yname = "disp",
             inc.density = TRUE)
plot(p1)
# plot with numerix x and factor y
p1 <- plottr(x = mtcars$mpg, y = as.factor(mtcars$cyl),
             xclass = "numeric", yclass = "factor",
             xname = "mpg", yname = "cyl", inc.density = TRUE)
plot(p1)

}
