% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_qqplot.R
\name{pval_qqplot}
\alias{pval_qqplot}
\title{Function for drawing a p-value qqplot}
\usage{
pval_qqplot(pvals, title = "QQ-plot")
}
\arguments{
\item{pvals}{The p-values you would like to look at using a QQ-plot}
}
\value{
A p-value QQ-plot
}
\description{
Function for drawing a p-value qqplot
}
\examples{
pval_qqplot(subcut_lm_results)
}
