% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_incomplete_covariates.R
\name{remove_incomplete_covariates}
\alias{remove_incomplete_covariates}
\title{Filter out individuals with incomplete phenotypes
Phenotypes filtered on: cohort, race, hardy, sex, ischemic time, age}
\usage{
remove_incomplete_covariates(with.mt.mito)
}
\description{
Filter out individuals with incomplete phenotypes
Phenotypes filtered on: cohort, race, hardy, sex, ischemic time, age
}
