% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_mito_outlier.R
\name{remove_mito_outlier}
\alias{remove_mito_outlier}
\title{Remove individuals that are outliers in corrected mito % (Greater or lower than 3 SD from the mean)
Greater than 3SD from the mean}
\usage{
remove_mito_outlier(with.mt.unalign)
}
\arguments{
\item{with.mt.unalign}{Data frame after filtering?}

\item{with_gtex}{data frame with GTEx IDs}
}
\value{
Transcripts for those individuals
}
\description{
Remove individuals that are outliers in corrected mito % (Greater or lower than 3 SD from the mean)
Greater than 3SD from the mean
}
