% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorg.R
\name{reorg}
\alias{reorg}
\title{Reorganizes the object to be in the right format, prints number of genes to look at}
\usage{
reorg(full)
}
\description{
Reorganizes the object to be in the right format, prints number of genes to look at
}
