% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lm.R
\name{run_lm}
\alias{run_lm}
\title{A wrapper for deciding either to run a default lm or a two-stage lm.  I always use the default.  Adapted from Vamsee (github.com/vkp3/pillalamarRi)}
\usage{
run_lm(expr, cov, SCORE, omit.outlier = T, method = "default")
}
\arguments{
\item{expr}{Expression vector (numeric) with length = #samples}

\item{cov}{Regression covariates in form [cov x samples]}

\item{SCORE}{The PRS, not included in `cov`}

\item{omit.outlier}{Whether or not you want to omit gene expression outliers}

\item{method}{Choose between 'default' or 'two-stage' for lm() method (see desc. in support functions below)}
}
\value{
A [1 x 8] vector output from an lm() like below:
        ['intercept', 'beta', 'SE', 't_value', 'pval', 'beta.conf.low', 'beta.conf.high', 'corr.rho']

Author: Vamsee Pillalamarri
}
\description{
formula: Expr ~ SCORE + Cov, where Cov = covariates
}
\details{
Runs linear regression to estimate effect of SCORE on gene expression
}
