% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_assoc2.R
\name{testing_assoc2}
\alias{testing_assoc2}
\title{Function that tests for associations and returns results.  DOES NOT merge with gene info, for use with tpm datasets.
This function will do a single linear regression and extract the coefficients from the regression.}
\usage{
testing_assoc2(lm_full,
  to_correct = "+ as.factor(covariates$SEX) + as.numeric(covariates$AGE)+ as.numeric(covariates$RACE) + covariates$PC1 + covariates$PC2 +covariates$PC3+ covariates$PC4 + covariates$PC5 + covariates$PC6 + covariates$PC7 + covariates$PC8 + covariates$PC9 + covariates$PC10")
}
\arguments{
\item{lm_full}{The data frame containing genes, covariates, and mtDNA-CN info}

\item{to_correct}{A string containing additional covariates to correct for.  ex: '+as.factor(covariates$DTHHRDY)+as.factor(covariates$RACE)'}

\item{lm_data}{Our dependent variables.  Column = dependent variable, rows = observations (in this case, individuals) (numeric vector)}
}
\value{
Coefficients from the regression
}
\description{
Function that tests for associations and returns results.  DOES NOT merge with gene info, for use with tpm datasets.
This function will do a single linear regression and extract the coefficients from the regression.
}
\examples{
lm_results <- testing_assoc2(lm_full)
}
