% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_amount_ngrams.R
\name{word_amount_ngrams}
\alias{word_amount_ngrams}
\title{count words and give ngrams}
\usage{
word_amount_ngrams(text, n = 1)
}
\arguments{
\item{text}{string to work with}

\item{n}{the int for ngrams, giving the number of words}
}
\value{
A list with a table of word frequencies and a vector of the ngrams
}
\description{
count words and give ngrams
}
\examples{
text <- "hey my name is andy what is your name"
word_amount_ngrams(text, n = 1)
}
