% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_word_frequency.R
\name{check_word_frequency}
\alias{check_word_frequency}
\title{use this function to get the count of a word of your word_amount_ngrams
output in Integer}
\usage{
check_word_frequency(summary, word)
}
\arguments{
\item{summary}{your word_amount_ngrams output}

\item{word}{the specific word that you want the frequency of}
}
\value{
A integer
}
\description{
use this function to get the count of a word of your word_amount_ngrams
output in Integer
}
\examples{
text <- "this day is a great day to talk about the next day"
summary <- word_amount_ngrams(text, 1)
check_word_frequency(summary, word = "day")
}
