% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{use this function to clean your String, by removing stopwords, changing it
to all lower case or remove its punctuation}
\usage{
clean_text(
  text,
  remove_stopwords = TRUE,
  to_lower = TRUE,
  remove_punctuation = TRUE
)
}
\arguments{
\item{text}{your String to clean}

\item{remove_stopwords}{a bool to set if you want to remove stopwords}

\item{to_lower}{a bool to set if you want to make to text to all lower case}

\item{remove_punctuation}{a bool to set if you want to remove punctuation}
}
\value{
A string of your cleaned text
}
\description{
use this function to clean your String, by removing stopwords, changing it
to all lower case or remove its punctuation
}
\examples{
text <- "Hey, my name is andy. What is your name?"
clean_text(text, remove_stopwords = FALSE, to_lower = TRUE, remove_punctuation = TRUE)
}
