% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_transcript.R
\name{examine_transcript}
\alias{examine_transcript}
\title{Gets the transcripts for people who remain after filtering}
\usage{
examine_transcript(lm_data, transcript, indiv = 304)
}
\arguments{
\item{counts1}{data frame containing the scaled counts from rse_gene (assay(rse_gene))}

\item{with_gtex}{data frame with GTEx IDs}
}
\value{
Transcripts for those individuals
}
\description{
Gets the transcripts for people who remain after filtering
}
\examples{
examine_transcript(blood_full, 'ENSG00000141527.16')
}
