% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateReport.R
\name{GenerateReport}
\alias{GenerateReport}
\title{Generate the report}
\usage{
GenerateReport(
  dtpath,
  catVars,
  yvar = NULL,
  model = "linReg",
  title = "Report",
  output_format = "html_document",
  tempDir = file.path(getwd(), "temp"),
  normality_test_method = "ks",
  interactive.plots = FALSE,
  include.vars = NULL
)
}
\arguments{
\item{dtpath}{dataset path}

\item{catVars}{vector of categorical variables names}

\item{yvar}{y variable name if present else \code{NULL}}

\item{model}{class of y variable (\code{numeric} or \code{factor}),
else \code{NULL}}

\item{title}{Title of the generated report}

\item{output_format}{output report format. \code{'html_documennt'} for
html file.}

\item{tempDir}{Directory where the output files needs to be stored.}

\item{normality_test_method}{method for normality test for a variable.
 Values can be \code{shapiro}
 for Shapiro-Wilk test or
 \code{'anderson'} for 'Anderson-Darling' test of normality or \code{ks} for
'Kolmogorov-Smirnov'}

\item{interactive.plots}{for interactive variable exploration}

\item{include.vars}{include only these variables from the full data}
}
\value{
creates a rmarkdown and html/pdf file
}
\description{
\code{GenerateReport} generates the markdown report in one command
}
\details{
This function cretes a rmarkdowwn report which can be converted to
html or pdf format file.
}
\examples{

\dontrun{
# Assigning the temporary folder in Documnets/temp fodler
GenerateReport(dtpath = "~/Documents/mtcars.csv", catVars = catVars,
               yvar = "cyl", model = "binClass",
               output_format = "html_document",
               title = "Report", tempDir = "~/Documents/temp",
               interactive.plots = FALSE)
}

}
