% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.for.cutoff.R
\name{perm.for.cutoff}
\alias{perm.for.cutoff}
\title{Function that permutes the dataset and finds an appropriate cutoff value.}
\usage{
perm.for.cutoff(permutations = 100, lm_full)
}
\arguments{
\item{permutations}{The number of times you want to permute the data}

\item{lm_full}{The dataset you want to test}
}
\value{
A 95% cutoff value, also a data frame with all the p-values from each permutation (to make the overlaid p-val qqplot)
}
\description{
Function that permutes the dataset and finds an appropriate cutoff value.
}
\examples{
subcut_lm_results <- apply(subcut_for_test, 2, lm_test, count = MT_count, covariates = covariates, correct_for = '+ as.factor(covariates$GENDER) + as.numeric(covariates$AGE)+ as.numeric(covariates$smrin) + as.factor(covariates$smcenter) + as.factor(covariates$RACE)+ as.numeric(covariates$TRISCHD) +as.factor(covariates$COHORT) + as.factor(covariates$DTHHRDY) + covariates$PC1 + covariates$PC2 +covariates$PC3+ covariates$PC4 + covariates$PC5')
}
