% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.analyzerPlot}
\alias{print.analyzerPlot}
\title{Method for print generic}
\usage{
\method{print}{analyzerPlot}(x, ...)
}
\arguments{
\item{x}{a plot of class \code{analyzerPlot}}

\item{...}{other parameters}
}
\value{
Displays the plot
}
\description{
This function plots the plot generated by the library \code{analyzer}
}
\examples{
# creating the plot
p <- plottr(mtcars$mpg)
p$x

}
