% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_assoc.R
\name{testing_assoc}
\alias{testing_assoc}
\title{Function that tests for associations and returns results
This function will do a single linear regression and extract the coefficients from the regression.}
\usage{
testing_assoc(lm_full)
}
\arguments{
\item{lm_data}{Our dependent variables.  Column = dependent variable, rows = observations (in this case, individuals) (numeric vector)}

\item{count}{Our independent variable (numeric vector)}

\item{covariates}{Covariates associated with each individual. Columns = covariates, rows = observations. (data.frame)}

\item{to_correct}{A string containing additional covariates to correct for.  ex: '+as.factor(covariates$DTHHRDY)+as.factor(covariates$RACE)'}
}
\value{
Coefficients from the regression
}
\description{
Function that tests for associations and returns results
This function will do a single linear regression and extract the coefficients from the regression.
}
\examples{
lm_results <- testing_assoc(lm_full)
}
