% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_amount_ngrams.R
\name{word_amount_ngrams}
\alias{word_amount_ngrams}
\title{Create a word count table and a ngrams list}
\usage{
word_amount_ngrams(text, n = 1)
}
\arguments{
\item{text}{string to work with}

\item{n}{the int for ngrams, giving the number of words}
}
\value{
A list with a table of word frequencies and a vector of the ngrams
}
\description{
create a table of all used words and their count, aswell as give a vector
of all the ngrams in the String where you define n (works best with clean
text)
}
\examples{
text <- "hey my name is andy what is your name"
word_amount_ngrams(text, n = 1)
}
